+++ MOUP/MOUP/Data/Network/AuthInterceptor.swift
@@ -9,10 +9,10 @@ import Foundation
 import Alamofire
 
 final class AuthInterceptor: RequestInterceptor {
-    private let authRepository: AuthRepositoryProtocol // TODO: - thread safety 추가 후 sendable 채택 필요
+    private let tokenRepository: TokenRepositoryProtocol // TODO: - thread safety 추가 후 sendable 채택 필요
     
-    init(authRepository: AuthRepositoryProtocol) {
-        self.authRepository = authRepository
+    init(tokenRepository: TokenRepositoryProtocol) {
+        self.tokenRepository = tokenRepository
     }
     
     func adapt(_ urlRequest: URLRequest, for session: Session, completion: @escaping (Result<URLRequest, any Error>) -> Void) {
@@ -34,7 +34,7 @@ final class AuthInterceptor: RequestInterceptor {
         // TODO: - 재발급 요청
         Task {
             do {
-                try await authRepository.renewAccessToken() // 액세스 토큰 재발급
+                try await tokenRepository.renewAccessToken() // 액세스 토큰 재발급
                 completion(.retry)
             } catch {
                 completion(.doNotRetryWithError(error))
