+++ MOUP/MOUP/Data/Services/AuthService.swift
@@ -11,7 +11,6 @@ import Alamofire
 protocol AuthServiceProtocol {
     func signIn(requestDTO: LoginRequestDTO) async throws -> LoginResponseDTO
     func signUp(requestDTO: RegisterRequestDTO) async throws -> RegisterResponseDTO
-    func renewAccessToken(requestDTO: RefreshTokenRequestDTO) async throws -> RefreshTokenResponseDTO
 }
 
 final class AuthService: AuthServiceProtocol {
@@ -64,29 +63,5 @@ final class AuthService: AuthServiceProtocol {
         }
     }
     
-    func renewAccessToken(requestDTO: RefreshTokenRequestDTO) async throws -> RefreshTokenResponseDTO {
-        let request = session.request(AuthRouter.renewAccessToken(requestDTO))
-        let response = await request.serializingDecodable(RefreshTokenResponseDTO.self).response
-        
-        guard let statusCode = response.response?.statusCode else {
-            throw NetworkError.noResponse
-        }
-        
-        switch statusCode {
-        case 200:
-            guard let dto = response.value else {
-                throw NetworkError.noResponse
-            }
-            return dto
-        case 400:
-            print(AuthError.invalidToken.debugDescription!)
-            throw AuthError.invalidToken
-        case 409:
-            print(AuthError.deletedUser.debugDescription!)
-            throw AuthError.deletedUser
-        default:
-            print(NetworkError.serverError.debugDescription!)
-            throw NetworkError.serverError
-        }
-    }
+    
 }
