+++ MOUP/MOUP/Domain/Interfaces/Repositories/AuthRepositoryProtocol.swift
@@ -10,5 +10,4 @@ import Foundation
 protocol AuthRepositoryProtocol: AnyObject {
     func signIn(requestDTO: LoginRequestDTO) async throws
     func signUp(requestDTO: RegisterRequestDTO) async throws -> UserRole
-    func renewAccessToken() async throws
 }
