+++ MOUP/MOUP/Domain/Interfaces/Repositories/TokenRepositoryProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  TokenRepositoryProtocol.swift
+//  MOUP
+//
+//  Created by 송규섭 on 10/17/25.
+//
+
+import Foundation
+
+protocol TokenRepositoryProtocol: AnyObject {
+    func renewAccessToken() async throws
+}
